Adding BotFramework support
BotFramework relies entirely on an async data path. A lot of refactoring had to be done to support async commands. If you use a callback base api in a bot command, you need to make sure that if you reply data to the chat that you use an async rapper and await on it, otherwise the BotFramework request will close and no longer be valid.